Add commission
+logout

+++ src/main/java/CarPackage/ClientRepository.java
@@ -10,9 +10,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ClientRepository extends CrudRepository<Client,Integer>{
+public interface ClientRepository extends CrudRepository<Client,Long>{
 
     List<Client> findAll();
+    //Client findOne(Long id);
 
     //@Override
     //@Query(value = "SELECT * FROM USERS WHERE EMAIL_ADDRESS = ?0", nativeQuery = true)
